Create tests

+++ src/main/java/entities/DigitalTelevisionContract.java
@@ -11,7 +11,7 @@ public class DigitalTelevisionContract extends DefaultContract{
     @Setter
     private int channelsCount;
 
-    public DigitalTelevisionContract(long ID, LocalDate startDate, LocalDate endDate, long number, Human owner, int channelsCount) {
+    public DigitalTelevisionContract(int ID, LocalDate startDate, LocalDate endDate, long number, Human owner, int channelsCount) {
         super(ID, startDate, endDate, number, owner);
         this.channelsCount = channelsCount;
     }
