+++ src/main/java/entities/WireInternetContract.java
@@ -11,7 +11,7 @@ public class WireInternetContract extends DefaultContract {
     @Setter
     private int speed;
 
-    public WireInternetContract(long ID, LocalDate startDate, LocalDate endDate, long number, Human owner, int speed) {
+    public WireInternetContract(int ID, LocalDate startDate, LocalDate endDate, long number, Human owner, int speed) {
         super(ID, startDate, endDate, number, owner);
         this.speed = speed;
     }
